update log output

+++ components/common/src/main/java/org/limewire/util/CommonUtils.java
@@ -285,29 +285,29 @@ public class CommonUtils {
      *     "h:mm:ss" where h=hours<24, mm=minutes, ss=seconds, or
      *     "m:ss" where m=minutes<60, ss=seconds
      */
-    public static String seconds2time(int seconds) {
-        int minutes = seconds / 60;
+    public static String seconds2time(long seconds) {
+        long minutes = seconds / 60;
         seconds = seconds - minutes * 60;
-        int hours = minutes / 60;
+        long hours = minutes / 60;
         minutes = minutes - hours * 60;
-        int days = hours / 24;
+        long days = hours / 24;
         hours = hours - days * 24;
         // build the numbers into a string
         StringBuilder time = new StringBuilder();
         if (days != 0) {
-            time.append(Integer.toString(days));
+            time.append(Long.toString(days));
             time.append(":");
             if (hours < 10) time.append("0");
         }
         if (days != 0 || hours != 0) {
-            time.append(Integer.toString(hours));
+            time.append(Long.toString(hours));
             time.append(":");
             if (minutes < 10) time.append("0");
         }
-        time.append(Integer.toString(minutes));
+        time.append(Long.toString(minutes));
         time.append(":");
         if (seconds < 10) time.append("0");
-        time.append(Integer.toString(seconds));
+        time.append(Long.toString(seconds));
         return time.toString();
     }
